Clarify `validUntil` description.
This clarifies the `validUntil` description to be more consistent. The
text claimed `validUntil` is the time the "credential ceases to be
valid", then that it is "the latest point in time [the info] is valid",
which can not both be true. The change here clarifies `validUntil` is
"the earliest point in time [the info] ceases to be valid".